requirements.txt
accelerate==1.6.0
aiohttp==3.11.16
anaconda-anon-usage==0.4.4
bitsandbytes==0.45.5
boto3==1.36.11
Brotli==1.0.9
conda-content-trust==0.2.0
coqui-tts==0.26.0
demucs==4.0.1
ftfy==6.3.1
gruut_lang_de==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
imblearn==0.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
lightgbm==4.6.0
nltk==3.9.1
openai-whisper==20240930
opencv-python==4.11.0.86
pathlib==1.0.1
pip==25.0
pyannote.audio==3.3.2
pyarrow==19.0.1
PySocks==1.7.1
python-dotenv==1.1.0
sacremoses==0.1.1
seaborn==0.13.2
SudachiDict-core==20250129
torchvision==0.21.0
wheel==0.45.1
xgboost==3.0.0
yt-dlp==2025.3.31
# LLM & API clients
mistralai
tqdm # For progress bars
TTS
